Initial listHeader component separation
